fix(removenode_filter_issue): the repair running
after `removenode` command is triggered generates a
message that is interpreted by SCT as an error, and
we try to ignore during the `disrupt_remove_node`
as one of the nodes is down. Filter was applied, but
there might be a delay between the executions, logging
and the SCT detecting these errors, so few tests were
still failing event with the filter in place. There
is still a need to add some extra grace period after
the "filter" is ended to not fail the test by late
logs.